third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-step/debugger-step-in-ignore-injected-script-expected.txt

CONSOLE MESSAGE: line 9: 123
Tests that stepInto doesn't pause in InjectedScriptSource.

Set timer for test function.
Call stack:
    0) testFunction (debugger-step-in-ignore-injected-script.html:10)
